Use the more general CharacterCategoryPatternExtractor instead of the  special case version that was in TimeFeaturesExtractors. (As an added bonus, performance went up a bit by doing so.)
git-svn-id: https://cleartk.googlecode.com/svn/trunk@2875 2f4ce1e4-9571-11dd-93b2-6d14b300b761

+++ cleartk-timeml/src/main/java/org/cleartk/timeml/eval/TempEval2010PipelineProvider.java
@@ -210,9 +210,11 @@ public class TempEval2010PipelineProvider extends CleartkPipelineProvider_ImplBa
     // if building to the pre-defined training directory, clean up non-model files
     for (CleartkInternalModelFactory factory : this.systemAnnotators.keySet()) {
       File modelDir = factory.getTrainingDirectory();
-      for (File file : modelDir.listFiles()) {
-        if (!file.isDirectory() && !file.getName().equals("model.jar")) {
-          file.delete();
+      if (modelDir.exists()) {
+        for (File file : modelDir.listFiles()) {
+          if (!file.isDirectory() && !file.getName().equals("model.jar")) {
+            file.delete();
+          }
         }
       }
     }
